+++ commands/stop.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+cd "$(dirname "$0")"
+
+pkill -f main.py
+rm -f running
+echo "Bot stopped..."