+++ commands/update.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+cd "$(dirname "$0")"
+
+./stop.sh
+git reset --hard
+git clean -xdf
+git pull
+
+chmod a+x *