+++ bot/cogs/matches.py
@@ -1,5 +1,6 @@
 from discord.ext import commands
 import discord
+from logging import getLogger
 
 from asyncio import sleep
 
@@ -15,8 +16,8 @@ from classes.maps import MapSelection
 from matches import getMatch, which_bot, which_team_channels
 from modules.enumerations import MatchStatus, SelStatus
 
-globId = 0
 
+log = getLogger(__name__)
 
 class MatchesCog(commands.Cog, name='matches'):
     """
@@ -28,8 +29,9 @@ class MatchesCog(commands.Cog, name='matches'):
 
     @commands.Cog.listener()
     async def on_ready(self):
-        print('Matches Cog is online')
-        return  # we don't display a message on each restart
+
+        log.info('Matches Cog is online')
+        return # we don't display a message on each restart
         try:
             for id in cfg.discord_ids["matches"]:
                 await channelSend("CHANNEL_INIT", id, id)