+++ bot/matches.py
@@ -13,6 +13,9 @@ from classes.accounts import AccountHander  # ok
 from random import choice as randomChoice
 from lib import tasks
 from asyncio import sleep
+from logging import getLogger
+
+log = getLogger(__name__)
 
 _lobbyList = list()
 _lobbyStuck = False
@@ -34,7 +37,15 @@ def addToLobby(player):
     _lobbyList.append(player)
     player.status = PlayerStatus.IS_LOBBIED
     if len(_lobbyList) == cfg.general["lobby_size"]:
+        _autoPing.cancel()
         startMatchFromFullLobby.start()
+    elif len(_lobbyList) >= cfg.general["lobby_size"] - 4:
+        if not _autoPing.is_running():
+            _autoPing.start()
+
+@tasks.loop(seconds=100, delay=1, count=2)
+async def _autoPing():
+    await channelSend("LB_NOTIFY", cfg.discord_ids["lobby"], f'<@&{cfg.discord_ids["notify_role"]}>')
 
 
 def getLobbyLen():
@@ -47,9 +58,8 @@ def getAllNamesInLobby():
 
 
 def removeFromLobby(player):
-    global _lobbyStuck
     _lobbyList.remove(player)
-    _lobbyStuck = False
+    _onLobbyRemove()
     player.status = PlayerStatus.IS_REGISTERED
 
 
@@ -57,6 +67,11 @@ def _onMatchFree():
     if len(_lobbyList) == cfg.general["lobby_size"]:
         startMatchFromFullLobby.start()
 
+def _onLobbyRemove():
+    global _lobbyStuck
+    _lobbyStuck = False
+    if len(_lobbyList) < cfg.general["lobby_size"]-4:
+        _autoPing.cancel()
 
 @tasks.loop(count=1)
 async def startMatchFromFullLobby():
@@ -64,8 +79,10 @@ async def startMatchFromFullLobby():
     match = _findSpotForMatch()
     if match == None:
         _lobbyStuck = True
+        _autoPing.cancel()
         await channelSend("LB_STUCK", cfg.discord_ids["lobby"])
         return
+    _lobbyStuck = False
     match._setPlayerList(_lobbyList)
     for p in _lobbyList:
         p.match = match  # Player status is modified automatically in IS_MATCHED
@@ -107,6 +124,7 @@ def clearLobby():
     for p in _lobbyList:
         p.status = PlayerStatus.IS_REGISTERED
     _lobbyList.clear()
+    _onLobbyRemove()
     return True
 
 
@@ -214,7 +232,8 @@ class Match:
 
     def confirmMap(self):
         self.__mapSelector.confirm()
-        self.__ready.start()
+        if self.__status == MatchStatus.IS_MAPPING:
+            self.__ready.start()
 
     def pickMap(self, captain):
         if self.__mapSelector.status == SelStatus.IS_SELECTED:
@@ -258,7 +277,7 @@ class Match:
             if p.hasOwnAccount:
                 continue
             if p.account == None:
-                print(f"Debug: {p.name} has no account")
+                log.error(f"Debug: {p.name} has no account")
             if p.account.isValidated:
                 continue
             notReady.append(p.mention)
@@ -290,7 +309,7 @@ class Match:
         for tm in self.__teams:
             tm.captain.isTurn = True
         if self.__mapSelector.status == SelStatus.IS_CONFIRMED:
-            await channelSend("MATCH_MAP_AUTO", self.__id, self.__map.name)
+            await channelSend("MATCH_MAP_AUTO", self.__id, self.__mapSelector.map.name)
             # ts3: map selected
             pick_channel = which_pick_channels(self.__id)
             ts3bot.move(pick_channel)
@@ -328,8 +347,10 @@ class Match:
         ts3.bot1.enqueue(ts3.AUDIO_ID_TYPE_READY)
         ts3.bot2.enqueue(ts3.AUDIO_ID_TYPE_READY)
 
+
     @tasks.loop(minutes=cfg.ROUND_LENGTH, delay=1, count=2)
     async def __onMatchOver(self):
+
         playerPings = [tm.allPings for tm in self.__teams]
         await channelSend("MATCH_ROUND_OVER", self.__id, *playerPings, self.roundNo)
         # ts3: round over
@@ -386,7 +407,7 @@ class Match:
         await channelSend("MATCH_STARTED", self.__id, *playerPings, self.roundNo)
         self.__roundsStamps.append(int(dt.timestamp(dt.now())))
         self.__status = MatchStatus.IS_PLAYING
-        self.__onMatchOver.start()
+        self._onMatchOver.start()
 
     @tasks.loop(count=1)
     async def _launch(self):
@@ -412,7 +433,7 @@ class Match:
         Team and ActivePlayer objects should get garbage collected, nothing is referencing them anymore"""
 
         if self.status == MatchStatus.IS_PLAYING:
-            self.__onMatchOver.cancel()
+            self._onMatchOver.cancel()
             playerPings = [tm.allPings for tm in self.__teams]
             await channelSend("MATCH_ROUND_OVER", self.__id, *playerPings, self.roundNo)
             # ts3: round over